Fixing bugs

+++ src/modules/core/services/admin/admin.service.ts
@@ -10,6 +10,7 @@ import { CreateAdminDto } from '../../dto/admin/create.admin';
 import { UpdateAdminDto } from '../../dto/admin/update.admin';
 import { UserService } from '../user/user.service';
 import { AdminException } from 'src/utils/exception';
+import { UserRequested } from 'src/infra/system/system.constant';
 
 @Injectable()
 export class AdminService {
@@ -40,6 +41,12 @@ export class AdminService {
   }
 
   async create(dto: CreateAdminDto): Promise<ResourceAdminDto> {
+    const authorized = await this.isAuthorized(UserRequested.userId);
+
+    if (!authorized) {
+      throw new AdminException('You are not authorized');
+    }
+
     const entity = await this._repo.create(new this._adminModel(dto));
     return this.getById(entity._id.toString());
   }
@@ -50,7 +57,17 @@ export class AdminService {
   }
 
   async delete(id: string): Promise<boolean> {
+    const authorized = await this.isAuthorized(UserRequested.userId);
+
+    if (!authorized) {
+      throw new AdminException('You are not authorized');
+    }
+
     const entity = await this.getById(id);
+
+    if (entity.status == '2') {
+      throw new AdminException('You can not delete active entities');
+    }
 
     if (entity.user.roles.length === 1) {
       await this._userService.delete(entity.user._id);
@@ -62,7 +79,51 @@ export class AdminService {
     return await this._repo.delete(id);
   }
 
+  async deactivate(id: string): Promise<ResourceAdminDto> {
+    const authorized = await this.isAuthorized(UserRequested.userId);
+
+    if (!authorized) {
+      throw new AdminException('You are not authorized');
+    }
+
+    const userRequested = await this._userService.getById(UserRequested.userId);
+    const dto = await this.getById(id);
+
+    if (dto.status == '1') {
+      throw new AdminException('This entity is still pending!');
+    }
+
+    const userEntity = await this.getByUserId(userRequested._id);
+    if (userEntity._id != id) {
+      throw new AdminException(
+        'You are not authorized to deactivate this entity.',
+      );
+    }
+
+    const entity = new UpdateAdminDto();
+    entity._id = dto._id;
+    entity.status = '3';
+    entity.user = dto.user;
+    return await this.update(entity);
+  }
+
+  private async isAuthorized(userId: string): Promise<boolean> {
+    const user = await this._userService.getById(userId);
+    let isAdmin = false;
+
+    if (user.roles.includes('0')) {
+      const admin = await this.getByUserId(userId);
+
+      if (admin && admin.status == '2') {
+        isAdmin = true;
+      }
+    }
+    return isAdmin;
+  }
+
   async getByUserId(id: string): Promise<ResourceAdminDto> {
+    await this.isAuthorized(UserRequested.userId);
+
     const entities = await this._repo.findAll();
     const entity = entities.find((entity) => entity.userId.toString() === id);
